Add first working version of ADCIRC-GSSHA coupling
The ADCIRC and GSSHA Python interfaces are dependencies of the
software. Their paths must be manually supplied in
watercoupler/watercoupler_path.py. Likewise, paths to the compiled
shared libraries of ADCIRC and GSSHA must be added to
LD_LIBRARY_PATH environment variable. Hopefully, though, if the
Python interfaces are pip-installed, we should eventually be able
to avoid having to add paths this way.

Four coupling types are available using the identifiers,
Adg, gdA, AdgdA, and gdAdg. The coupling is, in essence,
2D shallow water and 1D diffusive wave (which is in turn,
coupled to 2D diffusive wave).